requirements.txt
ultralytics==8.3.34
streamlit==1.32.2
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
